sensor controller add new sensor method has been refactored, but not tested

+++ SymfonyReact/src/Services/CardUserDataService.php
@@ -254,26 +254,19 @@ class CardUserDataService extends AbstractHomeAppUserSensorServiceCore
      * @return CardView
      * @throws Exception
      */
-    public function createNewSensorCard(Sensors $sensorObject): CardView
+    public function createNewSensorCard(Sensors $sensorObject): ?CardView
     {
         try {
-            $iconRepository = $this->em->getRepository(Icons::class);
-            $colourRepository = $this->em->getRepository(CardColour::class);
+            $randomIcon = $this->returnRandomIcon();
+
+            $randomColour = $this->returnRandomColour();
+
             $cardStateRepository = $this->em->getRepository(Cardstate::class);
 
-            $maxIconNumber = $iconRepository->countAllIcons();
-            $maxColourNumber = $colourRepository->countAllColours();
-
-
-            $firstIconId = $iconRepository->getFirstIconId()->getIconID();
-            $firstColourId = $colourRepository->getFirstColourId()->getColourID();
-
-            $randomIcon = $iconRepository->findOneBy(['iconID' => random_int($firstIconId, $firstIconId+$maxIconNumber-1)]);
-            $randomColour = $colourRepository->findOneBy(['colourID' => random_int($firstColourId, $maxColourNumber+$firstColourId-1)]);
             $onCardState = $cardStateRepository->findOneBy(['state' => Cardstate::ON]);
-//            dd($randomIcon, $randomColour, $onCardState, $sensorObject, $firstIconId);
-            if (!$randomIcon instanceof Icons && !$randomColour instanceof CardColour && !$onCardState instanceof Cardstate) {
-                throw new \RuntimeException('Something went wrong setting default values for card');
+
+            if (!$onCardState instanceof Cardstate) {
+                throw new \RuntimeException('Something went wrong setting a defualt card state');
             }
 
             $newCard = new CardView();
@@ -284,14 +277,48 @@ class CardUserDataService extends AbstractHomeAppUserSensorServiceCore
             $newCard->setCardStateID($onCardState);
 
             $this->em->persist($newCard);
-            $this->em->flush();
 
             return $newCard;
-        } catch (ORMException $e) {
-            $this->em->remove($newCard);
-            $this->em->remove($sensorObject);
+        } catch (\RuntimeException $exception) {
+            error_log($exception->getMessage());
+            $this->serverErrors[] = $exception->getMessage();
+        }
+        catch (ORMException $e) {
             error_log($e->getMessage());
             $this->serverErrors[] = 'Card Data Query Failure';
+        }
+
+        if (isset($newCard)) {
+            $this->em->remove($newCard);
+        }
+        $this->em->remove($sensorObject);
+
+        return null;
+    }
+
+    private function returnRandomIcon(): Icons
+    {
+        $iconRepository = $this->em->getRepository(Icons::class);
+        $maxIconNumber = $iconRepository->countAllIcons();
+        $firstIconId = $iconRepository->getFirstIconId()->getIconID();
+        $randomIcon = $iconRepository->findOneBy(['iconID' => random_int($firstIconId, $firstIconId+$maxIconNumber-1)]);
+
+        if (!$randomIcon instanceof Icons) {
+            throw new \RuntimeException('Failed setting random icon');
+        }
+
+        return $randomIcon;
+    }
+
+    private function returnRandomColour(): CardColour
+    {
+        $colourRepository = $this->em->getRepository(CardColour::class);
+        $maxColourNumber = $colourRepository->countAllColours();
+        $firstColourId = $colourRepository->getFirstColourId()->getColourID();
+        $randomColour = $colourRepository->findOneBy(['colourID' => random_int($firstColourId, $maxColourNumber+$firstColourId-1)]);
+
+        if (!$randomColour instanceof CardColour) {
+            throw new \RuntimeException('Failed setting random colour');
         }
     }
